Add and use `viewModelScope`, simplify a bit
We can use the `viewModelScope` extension from AndroidX to provide a ViewModel-bound coroutine scope instead of managing one ourselves.

We also remove a redundant LiveData (UI only needs one, and it’s better to limit the number of LiveData streams to prevent race conditions when updating the UI).

Finally, we remove the try/catch block for now, as there is no real use case for it.